+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix SQL importer type mappings: binary types, datetime/time, uuid now map to correct ODCS logicalType and format (#790)
 
 ### Added
+- Added support for MySQL for data contract tests (#1101)
 - Support additional PyArrow types in Parquet importer (#1091)
 - Populate `logicalTypeOptions.format` for SQL import from binary and uuid types (#790)
 - Snowflake DDL import with tags, descriptions, and template variable handling (#790)